requirements.txt
requests
aiohttp==3.8.1
asyncpg
pyaml